Let shared_ptrs cleanup CNodeStates
This removes reliance on net in net_processing for maintaining the
refcount == 0 invariant when calling FinalizeNode

This also removes a few asserts which used to be checked when there
were no more CNodeStates remaining, which we can no longer check for.

These should be pretty rarely checked anyway, so probably didn't
serve much use.