Make SyncTransaction provide an blocks instead of one transaction
This simplifies fixing #9148 as we can now hold cs_wallet across
an entire block instead of only per-tx.

This change also removes the NOT_IN_BLOCK constant in favor of only
passing the CBlockIndex* parameter to SyncTransactions when a new
block is being connected, instead of also when a block is being
disconnected.